10.12.2 fix 图片墙顺序不正确bug

+++ Metic/Main Classes/PictureWallViewController.h
@@ -24,8 +24,8 @@
 @property(nonatomic,strong)NSNumber *sequence;
 @property (nonatomic,strong) NSMutableArray* lefPhotos;
 @property (nonatomic,strong) NSMutableArray* rigPhotos;
-@property int leftH;
-@property int rightH;
+@property float leftH;
+@property float rightH;
 @property(nonatomic,strong)NSMutableArray *photo_list;//部分
 @property(nonatomic,strong)NSMutableArray *photo_list_all;//总
 @property(nonatomic,strong)NSMutableArray *photoPath_list;
